requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
soda-core-postgres==3.4.4
psycopg2-binary==2.9.10
pyyaml==6.0.2
gunicorn==23.0.0
python-dotenv==1.0.1
